pipeline-cli/openshift-client-x.test.ts rolled in lib.test.ts
